requirements.txt
streamlit>=1.32
anthropic>=0.40
pypdf>=4.2
python-docx>=1.1
python-dotenv>=1.0
